# src/inspect_ai/_view/inspect-openapi.json
{
  "components": {
    "schemas": {
...
      "EvalSet": {
        "properties": {
          "eval_set_id": {
            "title": "Eval Set Id",
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/EvalSetTask"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "eval_set_id",
          "tasks"
        ],
        "title": "EvalSet",
        "type": "object"
      },
...
      "LogDirResponse": {
        "properties": {
          "log_dir": {
            "title": "Log Dir",
            "type": "string"
          }
        },
        "required": [
          "log_dir"
        ],
        "title": "LogDirResponse",
        "type": "object"
      },
      "LogFilesResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/LogHandle"
            },
            "title": "Files",
            "type": "array"
          },
          "response_type": {
            "enum": [
              "incremental",
              "full"
            ],
            "title": "Response Type",
            "type": "string"
          }
        },
        "required": [
          "response_type",
          "files"
        ],
        "title": "LogFilesResponse",
        "type": "object"
      },
...
      "LogInfo": {
        "properties": {
          "direct_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direct Url"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "size"
        ],
        "title": "LogInfo",
        "type": "object"
      },
      "LogListingResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/LogHandle"
            },
            "title": "Files",
            "type": "array"
          },
          "log_dir": {
            "title": "Log Dir",
            "type": "string"
          }
        },
        "required": [
          "log_dir",
          "files"
        ],
        "title": "LogListingResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "inspect_ai types",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/eval-set": {
      "get": {
        "operationId": "eval_set_eval_set_get",
        "parameters": [
          {
            "in": "query",
            "name": "log_dir",
            "required": false,
            "schema": {
              "title": "Log Dir",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dir",
            "required": false,
            "schema": {
              "title": "Dir",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/EvalSet"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Eval Set Eval Set Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Eval Set"
      }
    },
...
    "/events": {
      "get": {
        "operationId": "api_events_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "last_eval_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Eval Time"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Api Events Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Events"
      }
    },
    "/flow": {
      "get": {
        "operationId": "flow_flow_get",
        "parameters": [
          {
            "in": "query",
            "name": "log_dir",
            "required": false,
            "schema": {
              "title": "Log Dir",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dir",
            "required": false,
            "schema": {
              "title": "Dir",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Flow"
      }
    },
    "/log-bytes/{log}": {
      "get": {
        "operationId": "api_log_bytes_log_bytes__log__get",
        "parameters": [
          {
            "in": "path",
            "name": "log",
            "required": true,
            "schema": {
              "title": "Log",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "title": "Start",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "title": "End",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Bytes"
      }
    },
    "/log-delete/{log}": {
      "get": {
        "operationId": "api_log_delete_log_delete__log__get",
        "parameters": [
          {
            "in": "path",
            "name": "log",
            "required": true,
            "schema": {
              "title": "Log",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Log Delete Log Delete  Log  Get",
                  "type": "boolean"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Delete"
      }
    },
    "/log-dir": {
      "get": {
        "operationId": "api_log_dir_log_dir_get",
        "parameters": [
          {
            "in": "query",
            "name": "log_dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Log Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogDirResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Dir"
      }
    },
    "/log-download/{log}": {
      "get": {
        "operationId": "api_log_download_log_download__log__get",
        "parameters": [
          {
            "in": "path",
            "name": "log",
            "required": true,
            "schema": {
              "title": "Log",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Download"
      }
    },
    "/log-files": {
      "get": {
        "operationId": "api_log_files_log_files_get",
        "parameters": [
          {
            "in": "query",
            "name": "log_dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Log Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogFilesResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Files"
      }
    },
    "/log-headers": {
      "get": {
        "operationId": "api_log_headers_log_headers_get",
        "parameters": [
          {
            "in": "query",
            "name": "file",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "File",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvalLog"
                  },
                  "title": "Response Api Log Headers Log Headers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Headers"
      }
    },
    "/log-info/{log}": {
      "get": {
        "operationId": "api_log_info_log_info__log__get",
        "parameters": [
          {
            "in": "path",
            "name": "log",
            "required": true,
            "schema": {
              "title": "Log",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogInfo"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Info"
      }
    },
    "/log-message": {
      "get": {
        "operationId": "api_log_message_log_message_get",
        "parameters": [
          {
            "in": "query",
            "name": "log_file",
            "required": true,
            "schema": {
              "title": "Log File",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "message",
            "required": true,
            "schema": {
              "title": "Message",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Message"
      }
    },
    "/log-size/{log}": {
      "get": {
        "operationId": "api_log_size_log_size__log__get",
        "parameters": [
          {
            "in": "path",
            "name": "log",
            "required": true,
            "schema": {
              "title": "Log",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Log Size Log Size  Log  Get",
                  "type": "integer"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log Size"
      }
    },
    "/logs": {
      "get": {
        "operationId": "api_logs_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "log_dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Log Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogListingResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Logs"
      }
    },
    "/logs/{log}": {
      "get": {
        "operationId": "api_log_logs__log__get",
        "parameters": [
          {
            "in": "path",
            "name": "log",
            "required": true,
            "schema": {
              "title": "Log",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "header-only",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Header-Only"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalLog"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Log"
      }
    },
...
    "/pending-samples": {
      "get": {
        "operationId": "api_pending_samples_pending_samples_get",
        "parameters": [
          {
            "in": "query",
            "name": "log",
            "required": true,
            "schema": {
              "title": "Log",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Samples"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api Pending Samples"
      }
    },
...
  }
}